moved js files to /public/js
