+++ public/js/starter_characters.json
